+++ glfw/cocoa_init.m
@@ -294,6 +294,18 @@ static GLFWbool initializeTIS(void)
 
 @end  // GLFWHelper
 
+@interface GLFWApplication : NSApplication
+@end
+
+extern void dispatchCustomEvent(NSEvent *event);
+
+@implementation GLFWApplication
+- (void)sendEvent:(NSEvent *)event {
+    if (event.type == NSEventTypeApplicationDefined) {
+        dispatchCustomEvent(event);
+    } else [super sendEvent:event];
+}
+@end
 
 //////////////////////////////////////////////////////////////////////////
 //////                       GLFW platform API                      //////
@@ -319,7 +331,7 @@ int _glfwPlatformInit(void)
                              toTarget:_glfw.ns.helper
                            withObject:nil];
 
-    [NSApplication sharedApplication];
+    [GLFWApplication sharedApplication];
 
     NSEvent* (^keydown_block)(NSEvent*) = ^ NSEvent* (NSEvent* event)
     {
@@ -375,6 +387,7 @@ int _glfwPlatformInit(void)
     if (!initializeTIS())
         return GLFW_FALSE;
 
+    _glfw.ns.displayLinks.lock = [NSLock new];
     _glfwInitTimerNS();
     _glfwInitJoysticksNS();
 
@@ -384,6 +397,12 @@ int _glfwPlatformInit(void)
 
 void _glfwPlatformTerminate(void)
 {
+    if (_glfw.ns.displayLinks.lock) {
+        _glfwClearDisplayLinks();
+        [_glfw.ns.displayLinks.lock release];
+        _glfw.ns.displayLinks.lock = nil;
+    }
+
     if (_glfw.ns.inputSource)
     {
         CFRelease(_glfw.ns.inputSource);